server/requirements.txt
flask
flask-cors
ftfy
pillow
regex
torch==1.7.1
torchvision
tqdm
